New roller working, begin to rewrite user info for the new sintaxe
