+++ data/io.github.kriptolix.Poliedros.metainfo.xml.in
@@ -57,14 +57,17 @@
   </screenshots>
 
   <releases>
-    <release version="1.1.0" date="2025-05-21">      
+    <release version="1.5.0" date="2025-05-21">      
       <description>
         <p>New functions and bug fixes.</p>
         <ul>
-          <li>"Count in" function added.</li>
+          <li>Count function added.</li>
+          <li>Reroll function added.</li>
           <li>Fudge dices roll added.</li>
           <li>d100 roll added.</li>
-          <li>Fixed some spelling errors.</li>          
+          <li>Nested functions added.</li>
+          <li>Some spelling errors fixed.</li>
+          <li>German translation added.</li>          
         </ul>
       </description>
     </release>
