+++ notes.md
@@ -1,17 +1,3 @@
-can
+corrigir log do explode para adicionar saida final
 
-keep subset -> reroll
-keep subset -> explode
-explode -> keep subset
-explode -> reroll
-reroll -> explode
-reroll -> keep subset
-
-5d6 + 1d4 | kh:3 | cn:>4 + 5d6 + | 1d4|kh:3 | cn:>4,
-
-
-dado = salva pool
-operação = soma pool adiciona operação
-
-operaçao com pool = 
-
+corrigir instruções de como usar comandos
