+++ src/tests/tests.py
@@ -1,7 +1,7 @@
-from ..commandparser import parse_expression
-from ..executor import execute_command 
+from ..roller import execute_command 
 
 tests = [
+    "4d8 + 1",
     "5d6+1d4|kh:3|cn:>4",
     "4d6 | kh:2 + 1",
     "d10",
@@ -10,7 +10,7 @@ tests = [
     "5d6+1|kl:1",    
     "4d6 | kl:2 + 5d10 |cn:>4",
     "3d6|ex:5,6 | cn:6",
-    "4d6|ex:5,6",
+    "4d6|ex:5..6",
     "3d6|rr:1,2",
 ]
 
@@ -23,11 +23,11 @@ def run_tests_all():
         if result[0]:
             print(f"{expression} -> {result}")
         else:
-            print(result[2])    
+            print(f"Fail with {result[2]}")    
 
 
 def run_tests_single():
-    expression = "5d6|rr:5,6"
+    expression = "5d6+1d4"
     # result = parse_expression(expression)
     result = execute_command(expression)
     print(f"{expression} -> {result}")
@@ -37,5 +37,5 @@ def run_tests_single():
 
 def run_tests():  # "4d6 + 1 | h:2 | c:5,6" 3d6|c:2
 
-    # run_tests_all()
-    run_tests_single()
+    run_tests_all()
+    # run_tests_single()
